refactor: extract HTTP server and STDIO handling to dedicated modules
- Move HTTP server logic from cmd/root.go to internal/http/server.go
- Extract JSON-RPC stdin handling to internal/stdio/stdio.go
- Add docs development server target to Makefile
- Enable system tray functionality with Show/Hide/Quit menu
- Remove config dependency from prompt tests for better isolation
- Clean up .gitignore entries and add planning/ directory
